+++ CHANGES.md
@@ -1,7 +1,9 @@
 2.1.0 (in development)
 ======================
 
-
+- Support for pickling `abc.abstractproperty`, `abc.abstractclassmethod`,
+  and `abc.abstractstaticmethod`.
+  ([PR #450](https://github.com/cloudpipe/cloudpickle/pull/450))
 
 2.0.0
 =====